mariadb: Insert in small chunks
Bulk inserts have an implicit transaction.
This creates a substantial undo/redo log,
which at some point has to be cleaned up.

As a side-effect there's random IO for about
the same duration as the original inserts,
and some queries may be slowed down by the
presence of a non-empty undo log.

Thus all benchmarks done during that time
have roughly half the IO available, and
substantial pressure on the innodb buffer pool.

Disable binlog while at it, it's only relevant
for replication and doesn't help us here.
(Which also reduces the amount of data written,
and on-disk size, by a substantial amount)