Apply feedback
Make the distinction between hot vs cold results more explicit. Do not
load file to the server upfront. Instead, in the first run make sure
that the file is not available as local file. Then, on the second/third
runs make sure that the file is locally available on the cluster.