Hyper: Fix EventTime handling in Parquet
The Parquet queries in Hyper were applying the `to_timestamp` function
to convert an integer to a timestamp, but this was incomplete because
that function takes the number of seconds since Unix epoch, and not the
number of days as stored in the EventDate column. To fix that, we now
use an explicit addition from the Unix epoch. This makes our results
slightly slower but correct.

Kudos to Mark Raasveldt for reporting the problem.